requirements.txt
# Core (lightweight) runtime dependencies for `pyimgano`.
#
# Optional backends (torch/onnx/openvino/etc.) are provided via extras:
#   pip install "pyimgano[torch]"
#   pip install "pyimgano[onnx]"
#   pip install "pyimgano[openvino]"
joblib>=1.0.0
numpy>=1.19
opencv-python>=4.5.0
Pillow>=8.0.0
scikit-learn>=0.22.0
scipy>=1.5.1
tqdm>=4.0.0
